requirements.txt
decorator==4.1.2
Django==2.2.2
django-leaflet==0.24.0
djangorestframework==3.9.4
httplib2==0.9.2
MarkupSafe==1.0
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
psycopg2-binary==2.8.3
PyJWT==1.5.3
pyodbc==4.0.26
django-pyodbc-azure==2.1.0.0
pyOpenSSL==17.5.0
pyserial==3.4
pytz==2019.1
requests==2.18.4
xlrd==1.2.0
